productivity_timer_agenda_v2_nobuild_sound_google/requirements.txt
plyer>=2.1.0
google-api-python-client>=2.131.0
google-auth>=2.35.0
google-auth-oauthlib>=1.2.0
